move interfaces -> types, remove redundant interfaces
