+++ frontend/src/components/pages/ecommerce/OrderCellContent.tsx
@@ -4,7 +4,7 @@ import Link from "next/link";
 import { useRouter } from "next/router";
 
 import { OrderFragment, PaymentStatus } from "@/generated/graphql";
-import { HeaderValuePair } from "@/interfaces/utils";
+import { HeaderValuePair } from "@/types/utils";
 
 type Props = { order: OrderFragment; field: HeaderValuePair<OrderFragment> };
