+++ frontend/src/components/pages/events/org/OrgEventsTable.tsx
@@ -3,7 +3,7 @@ import dayjs from "dayjs";
 import Link from "next/link";
 
 import { OrgAdminEventFragment, AdminOrganizationFragment } from "@/generated/graphql";
-import { HeaderValuePair } from "@/interfaces/utils";
+import { HeaderValuePair } from "@/types/utils";
 
 const eventFields: HeaderValuePair<OrgAdminEventFragment>[] = [
   { header: "Navn", field: "title" },
