--- frontend/src/interfaces/archive.ts
@@ -1,10 +1,0 @@
-export interface Document {
-  id: string;
-  title: string;
-  description: string;
-  date: string;
-  thumbnail: string;
-  typeDoc: string;
-  year: number | null;
-  webLink: string;
-}
