--- frontend/src/interfaces/organizations.ts
@@ -1,9 +1,0 @@
-export interface Organization {
-  id: string;
-  name: string;
-  description: string;
-  color: string | null;
-  children?: Organization[];
-  slug: string;
-  logoUrl?: string;
-}
