+++ frontend/src/pages/ecommerce/index.tsx
@@ -23,10 +23,10 @@ import { useRouter } from "next/router";
 
 import OrderCellContent from "@/components/pages/ecommerce/OrderCellContent";
 import { OrderFragment, UserDocument, UserOrdersDocument } from "@/generated/graphql";
-import { HeaderValuePair } from "@/interfaces/utils";
 import Layout, { RootStyle } from "@/layouts/Layout";
 import { addApolloState, initializeApollo } from "@/lib/apolloClient";
 import { NextPageWithLayout } from "@/pages/_app";
+import { HeaderValuePair } from "@/types/utils";
 
 const orderFields: HeaderValuePair<OrderFragment>[] = [
   { header: "Ordre-ID", field: "id" },
